requirements.txt
absl-py==0.7.1
astor==0.7.1
crowdai_api
cycler==0.10.0
dill==0.3.0
git+git://github.com/google-research/disentanglement_lib
enum34==1.1.6
future==0.17.1
gast==0.2.2
gin-config==0.1.4
grpcio==1.20.1
h5py==2.9.0
imageio==2.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
mock==3.0.4
numba==0.41.0
numpy==1.16.3
pandas==0.24.2
Pillow==6.2.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow==1.15.3
tensorflow-hub==0.4.0
tensorflow-probability==0.7.0
termcolor==1.1.0
Werkzeug==0.16.0
torch
torchvision
argparse
wandb
pytest
tqdm
pytest-dependency
matplotlib